Issue #404 - Save photos to Amazon S3

+++ familyconnections/gallery/index.php
@@ -524,8 +524,7 @@ Event.observe(window, \'load\', function() {
      */
     function displayDeletePhotoSubmit ()
     {
-        $photoId     = (int)$_POST['photo'];
-        $uploadsPath = getUploadsAbsolutePath();
+        $photoId = (int)$_POST['photo'];
 
         // Get photo info
         $sql = "SELECT `user`, `category`, `filename`, `external_id`
@@ -569,33 +568,29 @@ Event.observe(window, \'load\', function() {
         }
 
         // Remove any external photos for this photo
-        $sql = "DELETE FROM `fcms_gallery_external_photo`
-                WHERE `id` = ?";
-        if (!$this->fcmsDatabase->delete($sql, $photoExternalId))
+        if (!empty($photoExternalId))
         {
-            $this->displayHeader();
-            $this->fcmsError->displayError();
-            $this->displayFooter();
-            return;
+            $sql = "DELETE FROM `fcms_gallery_external_photo`
+                    WHERE `id` = ?";
+            if (!$this->fcmsDatabase->delete($sql, $photoExternalId))
+            {
+                $this->displayHeader();
+                $this->fcmsError->displayError();
+                $this->displayFooter();
+                return;
+            }
         }
 
-        $filePath  = $uploadsPath.'photos/member'.$photoUserId.'/'.basename($photoFilename);
-        $thumbPath = $uploadsPath.'photos/member'.$photoUserId.'/tb_'.basename($photoFilename);
-        $fullPath  = $uploadsPath.'photos/member'.$photoUserId.'/full_'.basename($photoFilename);
+        $this->Uploader = new Upload_PhotoGallery($this->fcmsError, $this->fcmsDatabase, $this->fcmsUser);
+
+        $filePath  = basename($photoFilename);
+        $thumbPath = 'tb_'.basename($photoFilename);
+        $fullPath  = 'full_'.basename($photoFilename);
 
         // Remove the Photo from the server
-        if (file_exists($filePath))
-        {
-            unlink($filePath);
-        }
-        if (file_exists($thumbPath))
-        {
-            unlink($thumbPath);
-        }
-        if (file_exists($fullPath))
-        {
-            unlink($fullPath);
-        }
+        $this->Uploader->deleteFile($filePath);
+        $this->Uploader->deleteFile($thumbPath);
+        $this->Uploader->deleteFile($fullPath);
 
         $_SESSION['message'] = 1;
